% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_extr.R
\name{code_extr}
\alias{code_extr}
\title{Code_extr}
\usage{
code_extr(level)
}
\arguments{
\item{level}{A character vector representing NUTS-level codes (e.g., \code{c("NUTS2", "NUTS3")}).}
}
\value{
A sorted numeric vector containing the extracted NUTS levels.
}
\description{
This function extracts the numerical value from NUTS-level strings.
}
