% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LAU.R
\name{get_LAU}
\alias{get_LAU}
\title{Get LAU data}
\usage{
get_LAU(year = "Null")
}
\arguments{
\item{year}{expressed as four digit (YYYY)}
}
\value{
A tibble containing  LAUs  information with selected columns (e.g., ISO, LAU_ID, NUTS3_ID  and geometry ).
}
\description{
Local Administrative Units (LAUs) are the building blocks of the NUTS classification and correspond to the municipalities and communes within the EU
To get the final dataframe we combine two dataset: one taken from Eurostat (https://ec.europa.eu/eurostat/web/nuts/local-administrative-units)that includes City names and City IDs, essential for querying and associations.
The other one taken from EEA which provides LAU information. The Latter dataset is updated automatically by selecting the most recent shapefile (SHP) available online.
While The Eurostat dataset URL needs to be manually updated with the latest download link to ensure the City-related data is current.
}
