% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NUTS.R
\name{get_NUTS}
\alias{get_NUTS}
\title{Get NUTS}
\usage{
get_NUTS(year = "Null")
}
\arguments{
\item{year}{expressed as four digit (YYYY)}
}
\value{
A tibble containing  LAUs  information with selected columns (NUTS_ID, LEVL_CODE...)
}
\description{
It automatically updates the dataset by identifying the most recent available file, accessing the corresponding page, and downloading the SHP file  at the 1:20 Million scale with the EPSG:4326 reference system from this website (https://gisco-services.ec.europa.eu/distribution/v2/nuts/)
}
