% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollutant.R
\name{get_pollutants}
\alias{get_pollutants}
\title{Get pollutant}
\usage{
get_pollutants()
}
\value{
A tibble containing pollutant information with selected columns (e.g., URI, notation, and extracted code).
}
\description{
Retrieve Pollutant Data from EEA Vocabulary (https://dd.eionet.europa.eu/vocabulary/aq/pollutant)
Downloads and processes pollutant data from the EEA (European Environment Agency) vocabulary database.
The data includes relevant information such as pollutant names, codes, and descriptions.
}
