% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_dates.R
\name{handle_dates}
\alias{handle_dates}
\title{Handle Dates based on Dataset Ranges}
\usage{
handle_dates(from, to)
}
\arguments{
\item{from}{StartDate (in "YYYY-MM-DD" format).}

\item{to}{EndDate (in "YYYY-MM-DD" format).}
}
\value{
A list of datasets with associated date ranges and descriptions.
}
\description{
This function handles dates based on the respective dataset. According to the documentation:
\itemize{
\item Data from 2024 onwards corresponds to Unverified data transmitted continuously (Up-To-Date/UTD/E2a).
\item Data from 2013 to the begin of 2023 corresponds to Verified data (E1a) reported by countries by 30 September each year for the previous year.
\item Data delivered before 2012 corresponds to Historical Airbase data.
The range for E1 is extended until 31/12/2023 because the observations are already validated, and no data for 2023 is retrieved when considering E2.
}
}
