% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEAaq_time_aggregate.R
\name{my_summarise}
\alias{my_summarise}
\title{Aggregate data based on a specific statistic}
\usage{
my_summarise(data, fun_aggr)
}
\arguments{
\item{data}{An \code{EEAaq_df} or \code{EEAaq_df_sfc} class object, which is the output of the
\code{\link{EEAaq_get_data}} function.}

\item{fun_aggr}{Vector character containing the aggregation function for which to time aggregate.}
}
\value{
A tibble with the required aggregation.
}
\description{
Given data and the aggregation function desired, this function compute a time aggregation of the data.
}
