% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollutants}
\alias{pollutants}
\title{Dataset containing information about the available pollutants}
\format{
A data frame with 650 rows and 4 variables:
\describe{
\item{Code}{Unique pollutant code identifier}
\item{Notation}{pollutant's short name}
\item{Label}{pollutant's full name}
\item{RecommendedUnit}{Measurement unit, recommended for the pollutant}
}
}
\source{
\url{http://dd.eionet.europa.eu/vocabulary/aq/pollutant/view}
}
\usage{
pollutants
}
\description{
This dataset belongs to the European Environment Agency institution. Contains short information about the available
pollutants.
}
\keyword{datasets}
