% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEAaq_get_data.R
\name{code_extr}
\alias{code_extr}
\title{Extract numeric NUTS code from the chracter}
\usage{
code_extr(level)
}
\arguments{
\item{level}{Character value of the NUTS level}
}
\value{
integer corresponding to the relative NUTS level or LAU.
}
\description{
This function take as argument the chracter NUTS or LAU code ("NUTS0", "NUTS1", "NUTS2", "NUTS3", "LAU")
and mutate it in a numerical hierarchical coding.
}
